added apis to get and post resume and photo
